pyproject.toml
[project]
name = "generate-gemini-voice"
version = "0.1.0"
description = "A modern CLI tool for generating voice using Google Cloud Text-to-Speech."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-cloud-texttospeech>=2.0.0",
    "pygame>=2.5.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.scripts]
generate-voice = "generate_gemini_voice.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/generate_gemini_voice"]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src --cov-report=term-missing"

[dependency-groups]
dev = [
    "pytest-cov>=7.0.0",
]
